refactor(js): extract export resolution pipeline
